Refactor inbox tests to its own file
